refactor: move SideNav to MobileNav component
