+++ polaris.shopify.com/src/components/MobileNav/MobileNav.module.scss
@@ -1,5 +1,7 @@
 @import "../../styles/variables.scss";
 
+$navItemPadding: 0.85rem 1rem;
+
 .MobileNav {
   display: flex;
   visibility: hidden;
@@ -28,6 +30,98 @@
         fill: var(--surface);
       }
     }
+  }
+}
+
+.LinksList {
+  display: flex;
+  align-items: flex-start;
+  flex-direction: column;
+  height: 100%;
+  width: calc(100vw - 4rem);
+  max-width: 22.5rem;
+  min-width: 15rem;
+  background: var(--surface);
+  font-size: var(--font-size-100);
+  font-weight: 500;
+  overflow-y: auto;
+  overflow-x: hidden;
+
+  li {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+
+    a {
+      width: 100%;
+      color: inherit;
+      padding: 0.85rem 1rem;
+
+      &:focus-within,
+      &:hover {
+        color: var(--text-strong);
+      }
+
+      &:hover,
+      &[aria-current="page"] {
+        background: rgba(0, 0, 0, 0.05);
+        border-radius: 0;
+      }
+    }
+  }
+}
+
+.AccordionHeader {
+  display: flex;
+  justify-content: space-between;
+  padding: $navItemPadding;
+  background-color: transparent;
+
+  svg {
+    transform: rotate(0deg);
+    transition: transform var(--duration-200) var(--ease-in-out);
+  }
+
+  &.expanded svg {
+    transform: rotate(90deg);
+  }
+
+  &:hover {
+    background: rgba(0, 0, 0, 0.05);
+    border-radius: 0;
+  }
+}
+
+.AccordionContent {
+  display: none;
+
+  li a {
+    padding-left: 2rem;
+  }
+
+  &.expanded {
+    display: block;
+  }
+}
+
+.Logo {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  padding: 0.85rem 1rem;
+  height: var(--header-height);
+  border-bottom: var(--border);
+  color: var(--text-inherit);
+  gap: 0.7rem;
+  font-weight: 500;
+  font-size: var(--font-size-100);
+
+  p {
+    margin: 0;
+  }
+
+  span {
+    transform: translate3d(0, -1px, 0);
   }
 }
