+++ polaris.shopify.com/src/components/MobileNav/MobileNav.tsx
@@ -1,11 +1,39 @@
 import { useState, useEffect, useRef } from "react";
+import Link from "next/link";
+import Image from "next/image";
 
-import { className } from "../../utils/various";
+import { foundationsNavItems, contributingNavItems } from "../../data/navItems";
+import { className, getComponentNav } from "../../utils/various";
 import { Breakpoints } from "../../types";
 import Button from "../Button";
-import SideNav from "../SideNav";
+import type { NavItem } from "../NavItems";
 
+import shopifyLogo from "../../../public/shopify-logo.svg";
 import styles from "./MobileNav.module.scss";
+
+const componentsNavItems = getComponentNav();
+
+const navItems: NavItem[] = [
+  {
+    title: "Getting started",
+    url: "/resources",
+  },
+  ...foundationsNavItems,
+  {
+    title: "Components",
+    url: "/components",
+    children: componentsNavItems[0].children,
+  },
+  {
+    title: "Tokens",
+    url: "/tokens/colors",
+  },
+  {
+    title: "Icons",
+    url: "/icons",
+  },
+  ...contributingNavItems,
+];
 
 interface Props {
   currentPath: string;
@@ -14,7 +42,8 @@ interface Props {
 function MobileNav({ currentPath }: Props) {
   const [showMenu, setShowMenu] = useState(false);
   const menuButtonRef = useRef<HTMLButtonElement>(null);
-  const menuNavRef = useRef<HTMLElement>(null);
+  const logoRef = useRef<HTMLAnchorElement>(null);
+  const closeButtonRef = useRef<HTMLButtonElement>(null);
 
   useEffect(() => {
     function hideSideNavOnResize() {
@@ -28,6 +57,21 @@ function MobileNav({ currentPath }: Props) {
     return () => window.removeEventListener("resize", hideSideNavOnResize);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  useEffect(() => {
+    const handleOnKeyDown = (e: KeyboardEvent) => {
+      if (e.key === "Escape") {
+        setShowMenu(false);
+      }
+    };
+
+    if (showMenu) {
+      document.addEventListener("keydown", handleOnKeyDown);
+      logoRef.current instanceof HTMLElement && logoRef.current.focus();
+    }
+
+    return () => document.removeEventListener("keydown", handleOnKeyDown);
+  }, [showMenu]);
 
   const handleCloseMenu = () => {
     setShowMenu(false);
@@ -52,27 +96,54 @@ function MobileNav({ currentPath }: Props) {
       <nav
         id="side-menu"
         className={className(styles.MobileNav, showMenu && styles.show)}
-        ref={menuNavRef}
       >
-        <SideNav
-          currentPath={currentPath}
-          showMenu={showMenu}
-          handleCloseMenu={handleCloseMenu}
-        />
+        <ul className={styles.LinksList}>
+          <li>
+            <Link href="/" passHref>
+              <a
+                ref={logoRef}
+                className={styles.Logo}
+                onClick={handleCloseMenu}
+                onKeyDown={(e) => {
+                  if (e.key === "Tab" && e.shiftKey) {
+                    e.preventDefault();
+                    const closeButton = closeButtonRef.current;
+                    closeButton instanceof HTMLElement && closeButton.focus();
+                  }
+                }}
+              >
+                <Image
+                  src={shopifyLogo}
+                  layout="fixed"
+                  width={24}
+                  height={24}
+                  alt="Shopify logo"
+                />
+                Polaris
+              </a>
+            </Link>
+          </li>
+
+          {navItems.map((item, idx) => (
+            <ListItem
+              key={idx}
+              item={item}
+              currentPath={currentPath}
+              handleCloseMenu={handleCloseMenu}
+            />
+          ))}
+        </ul>
 
         <button
+          ref={closeButtonRef}
           aria-label="Close menu"
           className={styles.CloseButton}
           onClick={handleCloseMenu}
           onKeyDown={(e) => {
             if (e.key === "Tab" && !e.shiftKey) {
               e.preventDefault();
-              // TODO: create a direct ref to the Logo
-              const polarisLogoLink =
-                menuNavRef.current?.firstElementChild?.children[0]
-                  .firstElementChild;
-
-              polarisLogoLink instanceof HTMLElement && polarisLogoLink.focus();
+              const logoLink = logoRef.current;
+              logoLink instanceof HTMLElement && logoLink.focus();
             }
           }}
         >
@@ -84,6 +155,69 @@ function MobileNav({ currentPath }: Props) {
         <div className={styles.Backdrop} onClick={handleCloseMenu} />
       )}
     </>
+  );
+}
+
+interface ListItemProps {
+  item: NavItem;
+  currentPath: string;
+  handleCloseMenu?: () => void;
+}
+
+function ListItem({ item, currentPath, handleCloseMenu }: ListItemProps) {
+  const [isExpanded, setIsExpanded] = useState(false);
+
+  return (
+    <li>
+      {item.children && (
+        <>
+          <button
+            className={className(
+              styles.AccordionHeader,
+              isExpanded && styles.expanded
+            )}
+            onClick={() => setIsExpanded((prevState) => !prevState)}
+          >
+            {item.title} <ChevronRightIcon />
+          </button>
+
+          <ul
+            className={className(
+              styles.AccordionContent,
+              isExpanded && styles.expanded
+            )}
+          >
+            {item.children.map((child) => (
+              <ListItem
+                key={child.url}
+                item={child}
+                currentPath={currentPath}
+                handleCloseMenu={handleCloseMenu}
+              />
+            ))}
+          </ul>
+        </>
+      )}
+
+      {!item.children && item.url && (
+        <Link href={item.url} passHref>
+          <a
+            aria-current={item.url === currentPath ? "page" : "false"}
+            onClick={handleCloseMenu}
+          >
+            {item.title}
+          </a>
+        </Link>
+      )}
+    </li>
+  );
+}
+
+function ChevronRightIcon() {
+  return (
+    <svg viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+      <path d="M8 16a.999.999 0 0 1-.707-1.707l4.293-4.293-4.293-4.293a.999.999 0 1 1 1.414-1.414l5 5a.999.999 0 0 1 0 1.414l-5 5a.997.997 0 0 1-.707.293z" />
+    </svg>
   );
 }
