--- polaris.shopify.com/src/components/SideNav/index.ts
@@ -1,3 +1,0 @@
-import SideNav from "./SideNav";
-
-export default SideNav;
